modified api

+++ backend/controllers/userController.js
@@ -38,30 +38,32 @@ exports.registerUser = catchAsyncErrors(async (req, res, next) => {
 });
 
 //Register Mobile User
-exports.registerMobile = catchAsyncErrors(async (req, res) => {
+exports.registerMobile = async (req, res) => {
+  try{
     const { name, email, password } = req.body;
-    console.log(req.files.avatar.tempFilePath)
+    
+    const avatar = req.files.avatar.tempFilePath;
+    // console.log(req.files.avatar.tempFilePath)  
+      const mycloud = await cloudinary.v2.uploader.upload(avatar, {
+        folder: "avatars",
+        width: 150,
+        crop: "scale",
+      });
+      // console.log(mycloud)
+    let user = await User.findOne({ email });
 
-    const avatar = req.files.avatar.tempFilePath;  
-    // let user = await User.findOne({ email });
-
-    // if (user) {
-    //   return res
-    //     .status(400)
-    //     .json({ success: false, message: "User already exists" });
-    // }
+    if (user) {
+      return res
+        .status(400)
+        .json({ success: false, message: "User already exists" });
+    }
 
 
-    const mycloud = await cloudinary.v2.uploader.upload(avatar, {
-      folder: "avatars",
-      width: 150,
-      crop: "scale",
-    });
 
     fs.rmSync("./tmp", { recursive: true });
     
 
-   const user = await User.create({
+    user = await User.create({
       name,
       email,
       password,
@@ -73,8 +75,11 @@ exports.registerMobile = catchAsyncErrors(async (req, res) => {
 
     sendToken(user, 201, res);
     
-  } 
-)
+  } catch(error){
+    res.status(500).json({ success: false, message: error.message });
+  }
+}
+
 
 //Login User
 exports.loginUser = catchAsyncErrors(async (req, res, next) => {
